management/requirements.txt
Django==4.1.2
mysqlclient==2.1.1
python-prctl==1.8.1